DGJ 673 enable release not popup on all page after successful login
