+++ forms-flow-web/src/containers/styles.scss
@@ -412,6 +412,7 @@
 
 .menu-padded {
   padding: 5px 0 5px 0;
+  font-size: 1rem;
 }
 
 .margin-bottom-10 {
